+++ cronsetup/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## To be Released
 
+* Update Scalingo/go-etcd-cron dep and bump etcd client to 3.5.0
+  [#207](https://github.com/Scalingo/go-utils/pull/207)
 * Bump go version to 1.16
 
 ## v1.0.0, v1.0.1, v1.0.2